requirements.txt
aiogram==3.3.0
fastapi==0.109.1
uvicorn==0.27.0
aiosqlite==0.19.0
python-dotenv==1.0.0
jinja2==3.1.3
python-multipart==0.0.18
aiofiles==23.2.1
apscheduler==3.10.4
passlib==1.7.4
bcrypt==4.1.2
